+++ test/test_expression.jl
@@ -1118,7 +1118,7 @@ end
 
         alpine = _build(m)
 
-        @test alpine.num_constr_convex == 21
+        @test alpine.num_constr_convex == 19
 
         # 0 : OBJ
         @test alpine.obj_structure == :convex
@@ -1181,29 +1181,7 @@ end
         @test alpine.bounding_constr_mip[5][:powers] == [2, 2, 2]
         @test alpine.bounding_constr_mip[5][:cnt] == 3
 
-        # 6
-        @test alpine.constr_structure[6] == :convex
-        @test alpine.nonlinear_constrs[6][:expr_orig] == :constraints
-        @test alpine.nonlinear_constrs[6][:convex_type] == :convexC
-        @test alpine.nonlinear_constrs[6][:convexified] == :false
-        @test alpine.bounding_constr_mip[6][:sense] == :(<=)
-        @test alpine.bounding_constr_mip[6][:coefs] == [3.0, 4.0, 5.0]
-        @test alpine.bounding_constr_mip[6][:vars] == [:(x[1]), :(x[2]), :(x[5])]
-        @test alpine.bounding_constr_mip[6][:rhs] == 100.0
-        @test alpine.bounding_constr_mip[6][:powers] == [0.5, 0.5, 0.5]
-        @test alpine.bounding_constr_mip[6][:cnt] == 3
-
-        # 7
-        @test alpine.constr_structure[7] == :convex
-        @test alpine.nonlinear_constrs[7][:expr_orig] == :constraints
-        @test alpine.nonlinear_constrs[7][:convex_type] == :convexC
-        @test alpine.nonlinear_constrs[7][:convexified] == :false
-        @test alpine.bounding_constr_mip[7][:sense] == :(>=)
-        @test alpine.bounding_constr_mip[7][:coefs] == [-3.0, -4.0]
-        @test alpine.bounding_constr_mip[7][:vars] == [:(x[1]), :(x[2])]
-        @test alpine.bounding_constr_mip[7][:rhs] == -100.0
-        @test alpine.bounding_constr_mip[7][:powers] == [0.5, 0.5]
-        @test alpine.bounding_constr_mip[7][:cnt] == 2
+        #6,7 excluded 
 
         # 8
         @test alpine.constr_structure[8] == :convex
@@ -3661,4 +3639,58 @@ end
     @test isapprox(JuMP.objective_value(m), 2.0; atol = 1E-6)
     @test isapprox(JuMP.value(z), 1; atol = 1e-6)
     @test isapprox(JuMP.value(x), 1; atol = 1e-6)
+end
+
+@testset "@NLexpression from quadratic @expression (Issue #221)" begin
+    @testset "Unit tests for `expr_is_emptysum`" begin
+        # Empty sum
+        @test Alpine.expr_is_emptysum(:(+()))
+        # Not empty sum zero int
+        @test !Alpine.expr_is_emptysum(:(+(0, 0)))
+        # Not empty sum zero float
+        @test !Alpine.expr_is_emptysum(:(+(0.0, 0.0)))
+        # Not empty sum int with zero
+        @test !Alpine.expr_is_emptysum(:(+(1, 0)))
+        # Not empty sum int
+        @test !Alpine.expr_is_emptysum(:(+(1, 2)))
+        # Not empty sum float with zero
+        @test !Alpine.expr_is_emptysum(:(+(1.0, 0.0)))
+        # Not empty sum float
+        @test !Alpine.expr_is_emptysum(:(+(1.0, 3.0)))
+        # Not empty sum of variables
+        @test !Alpine.expr_is_emptysum(:(x + y))
+        # Not empty sum of variable with int
+        @test !Alpine.expr_is_emptysum(:(x + 1))
+        # Not empty sum of variable with float
+        @test !Alpine.expr_is_emptysum(:(x + 3.0))
+        # Not empty expression with non-linear function evaluation
+        @test !Alpine.expr_is_emptysum(:(sin(x)))
+        # Not empty expression with anonimus function
+        @test !Alpine.expr_is_emptysum(:(x -> log(x)))
+        # Not empty expression with anonimus function
+        @test !Alpine.expr_is_emptysum(:(x -> log(x)))
+        # Not empty expression with empty sum
+        @test !Alpine.expr_is_emptysum(:($(:(+())) + 1))
+    end
+
+    test_solver = JuMP.optimizer_with_attributes(
+        Alpine.Optimizer,
+        "nlp_solver" => IPOPT,
+        "mip_solver" => PAVITO,
+        "presolve_bt" => true,
+        "apply_partitioning" => true,
+    )
+
+    m = JuMP.Model(test_solver)
+
+    @variable(m, -2 ≤ x ≤ 2, start = 1)
+    @expression(m, expr, x^2)
+    @NLconstraint(m, expr >= 1)
+    @NLconstraint(m, expr <= 2)
+    @NLobjective(m, Min, x^3)
+
+    JuMP.optimize!(m)
+    @test JuMP.termination_status(m) == JuMP.MOI.OPTIMAL
+    @test isapprox(JuMP.objective_value(m), -(sqrt(2))^3; atol = 1E-6)
+    @test isapprox(JuMP.value(m[:x]), -sqrt(2); atol = 1E-6)
 end